# infra/api-gateway/openapi.yaml
openapi: "3.0.1"
info:
  title: "pedidos360-api"
  description: >
    Definicion del AWS API Gateway (HTTP API) para Pedidos360.
    Punto de entrada unico hacia los microservicios de Productos y Carrito.
    Valida el JWT de Azure AD en el borde (firma via JWKS + issuer + audience).
  version: "1.0.0"

# CORS: solo el dominio del frontend puede llamar al gateway.
...
components:
  securitySchemes:
    azureJwt:
      type: oauth2
      flows: {}
      # Authorizer nativo de tipo JWT del HTTP API.
      x-amazon-apigateway-authorizer:
        type: jwt
        identitySource: "$request.header.Authorization"
        jwtConfiguration:
          # issuer = https://login.microsoftonline.com/<TENANT_ID>/v2.0
          issuer: "https://login.microsoftonline.com/e5372bf0-c5e3-4286-887c-79069f209c1f/v2.0"
          # audience = App ID URI o Client ID de la API (Azure v2 puede emitir cualquiera)
          audience:
            - "api://ce3e5b61-0b01-4a67-a42b-f15ccb364024"
            - "ce3e5b61-0b01-4a67-a42b-f15ccb364024"

# Por defecto toda ruta exige un JWT valido; el catalogo (GET) se marca publico.
security:
  - azureJwt: []
paths:
...
  /productos/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Detalle de producto (publico)
      security: []
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: GET
        uri: "http://3.225.255.55:8081/api/productos/{id}"
        connectionType: INTERNET
        payloadFormatVersion: "1.0"
    put:
      summary: Actualizar producto
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: PUT
        uri: "http://3.225.255.55:8081/api/productos/{id}"
        connectionType: INTERNET
        payloadFormatVersion: "1.0"
    delete:
      summary: Eliminar producto (rol ADMIN)
      x-amazon-apigateway-integration:
        type: http_proxy
        httpMethod: DELETE
        uri: "http://3.225.255.55:8081/api/productos/{id}"
        connectionType: INTERNET
        payloadFormatVersion: "1.0"

  # =========================================================
  #  Carrito / Pedidos  ->  EC2 :8082  /api/carrito   (todo protegido)
  # =========================================================
...